Implemented #300
