+++ package.json
@@ -44,12 +44,20 @@
     "gulp-rename": "^1.2.2",
     "gulp-sourcemaps": "^1.6.0",
     "gulp-tslint": "^6.0.1",
-    "gulp-typedoc": "^2.0.0",
     "gulp-typescript": "^2.13.0",
     "gulp-uglify": "^1.5.1",
     "harmonize": "^1.4.4",
     "harmony-proxy": "^1.0.0",
     "istanbul": "^0.4.2",
+    "karma": "^1.1.2",
+    "karma-chai": "^0.1.0",
+    "karma-chrome-launcher": "^1.0.1",
+    "karma-commonjs": "^1.0.0",
+    "karma-firefox-launcher": "^1.0.0",
+    "karma-ie-launcher": "^1.0.0",
+    "karma-mocha": "^1.1.1",
+    "karma-mocha-reporter": "^2.0.5",
+    "karma-sinon": "^1.0.5",
     "mocha": "^2.5.1",
     "publish-please": "^2.1.3",
     "reflect-metadata": "^0.1.3",
